shader optimizations, ps5 fixes

+++ WickedEngine/shaders/ShaderInterop_VoxelGrid.h
@@ -56,8 +56,8 @@ struct ShaderVoxelGrid
 		sub_coord.y = coord.y % 4u;
 		sub_coord.z = coord.z % 4u;
 		const uint bit = flatten3D(sub_coord, uint3(4, 4, 4));
-		const uint64_t mask = 1ull << bit;
-		return (voxels_4x4_block & mask) != 0ull;
+		const uint64_t mask = uint64_t(1) << bit;
+		return (voxels_4x4_block & mask) != uint64_t(0);
 	}
 	bool check_voxel(float3 worldpos)
 	{
